Just annoying things

+++ src/main/java/sardorcreate/dto/commutator/CommDto.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.*;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 @Getter
@@ -19,7 +19,7 @@ public class CommDto {
     private String model;
     private LocalDate date;
     private String whereFrom;
-    private BigDecimal price;
+    private long price;
     private ToolsStatus status;
     private CommutatorType type;
     private PortType portType;
