+++ src/main/java/sardorcreate/dto/printer/PrinterCreateDto.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.PrinterType;
 
-import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -15,7 +15,7 @@ public class PrinterCreateDto {
     private long inventoryId;
     private String model;
     private String whereFrom;
-    private BigDecimal price;
+    private long price;
     private PrinterType type;
     private boolean isColored;
     private int colorCount;
