+++ src/main/java/sardorcreate/dto/tablet/TabletDto.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 import sardorcreate.enums.OSType;
 import sardorcreate.enums.ToolsStatus;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 @Getter
@@ -20,7 +20,7 @@ public class TabletDto {
     private String model;
     private LocalDate date;
     private String whereFrom;
-    private BigDecimal price;
+    private long price;
     private ToolsStatus status;
     private OSType osType;
     private int ram;
